Revert "CU-8697xj2u4 refactor: broke up useSignupHandler in parts"
